+++ docs/more_info/changelog.rst
@@ -7,7 +7,7 @@ Change Log
 Version:
 ================
 
-- New feature: upon reading mge data, q=1 values are set to q=0.99999 for numerical stability
+- New feature: upon reading mge data, q values too close to 1 are set to q=0.99999 for numerical stability
 - New feature: the number of configuration file backups can be better controlled by ``config_reader.Configuration.backup_config_file(...)`` options
 - Improvement: The bash test script ``dev_tests/test_all.sh`` executes a grid of test scenarios (different base scripts with different parameter generators and weight solvers) either locally or via Slurm
 - Improvement: Greatly improved performance of the chi2 plot